requirements.txt
telethon==1.33.1
pyyaml==6.0.1
schedule==1.2.1
croniter==2.0.1
python-crontab==3.0.0
